backend/pyproject.toml
[project]
name = "tunify-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.3",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "psycopg2-binary>=2.9.11",
    "asyncpg>=0.30.0",
    "alembic>=1.17.0",
    "bcrypt==4.0.1",
    "passlib==1.7.4",
    "python-multipart>=0.0.20",
    "python-jose[cryptography]>=3.5.0",
]
